% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load.r
\name{load.repo}
\alias{load.repo}
\title{Load package from local git repository folder.}
\usage{
load.repo(name, folder = "package")
}
\arguments{
\item{name}{A package name for loading.}

\item{folder}{A folder with packages inside the repository.}
}
\description{
Load package from local git repository folder.
}
\details{
Function assumes that R working folder is inside git repository.
}

